pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "radslice"
version = "0.1.0"
description = "Multimodal radiology LLM benchmark across CT, MRI, X-ray, and Ultrasound"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "GOATnote Inc." }]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-genai>=1.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
    "coverage>=7.0",
]

[project.scripts]
radslice = "radslice.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"src/radslice/grading/judge.py" = ["E501"]
"tests/test_grading.py" = ["E501"]
"tests/test_smoke.py" = ["E501", "E741"]
"tests/test_transcript.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "smoke: end-to-end smoke tests",
]
